+++ packages/spaces/commands/vpn/destroy.js
@@ -3,15 +3,23 @@
 const cli = require('heroku-cli-util')
 const co = require('co')
 
-function * run (context, heroku) {
-  let space = context.flags.space || context.args.space
-  if (!space) throw new Error('Space name required.\nUSAGE: heroku spaces:vpn:destroy --space my-space')
+function check (val, message) {
+  if (!val) throw new Error(`${message}.\nUSAGE: heroku spaces:vpn:destroy --space example-space vpn-connection-name`)
+}
 
-  let lib = require('../../lib/vpn')(heroku)
+function * run (context, heroku) {
+  let space = context.flags.space
+  check(space, 'Space name required')
+
+  let name = context.flags.name || context.args.name
+  // TODO: uncomment when we've fully migrated to the multiple VPN UX
+  // check(name, 'VPN name required')
+
+  let lib = require('../../lib/vpn-connections')(heroku)
 
   yield cli.confirmApp(space, context.flags.confirm, `Destructive Action
-This command will attempt to destroy VPN in space ${cli.color.green(space)}`)
-  yield cli.action(`Tearing down VPN in space ${cli.color.cyan(space)}`, lib.deleteVPN(space))
+This command will attempt to destroy the specified VPN Connection in space ${cli.color.green(space)}`)
+  yield cli.action(`Tearing down VPN Connection in space ${cli.color.cyan(space)}`, lib.deleteVPNConnection(space, name))
 }
 
 module.exports = {
@@ -20,15 +28,18 @@ module.exports = {
   description: 'destroys VPN in a private space',
   help: `Example:
 
-    $ heroku spaces:vpn:destroy --confirm --space my-space
-    Tearing down VPN in space my-space
+    $ heroku spaces:vpn:destroy --confirm --space example-space vpn-connection-name
+    Tearing down VPN Connection in space example-space
   `,
   hidden: true,
   needsApp: false,
   needsAuth: true,
-  args: [{name: 'space', optional: true, hidden: true}],
+  args: [
+    {name: 'name', optional: true, hidden: true}
+  ],
   flags: [
     {name: 'space', char: 's', hasValue: true, description: 'space to get peering info from'},
+    {name: 'name', char: 'n', hasValue: true, description: 'name or id of the VPN connection to retrieve config from'},
     {name: 'confirm', hasValue: true, description: 'set to space name bypass confirm prompt'}
   ],
   run: cli.command(co.wrap(run))