+++ packages/spaces/test/commands/vpn/destroy.js
@@ -9,13 +9,23 @@ const cli = require('heroku-cli-util')
 describe('spaces:vpn:destroy', function () {
   beforeEach(() => cli.mockConsole())
 
-  it('destroys VPN', function () {
+  it('destroys VPN Connection when name is specified', function () {
+    let api = nock('https://api.heroku.com:443')
+      .delete('/spaces/my-space/vpn-connections/my-vpn-connection')
+      .reply(202)
+    return cmd.run({args: {name: 'my-vpn-connection'}, flags: {space: 'my-space', confirm: 'my-space'}})
+      .then(() => expect(cli.stderr).to.equal(
+        `Tearing down VPN Connection in space my-space... done\n`))
+      .then(() => api.done())
+  })
+
+  it('destroys VPN when no name is specified', function () {
     let api = nock('https://api.heroku.com:443')
       .delete('/spaces/my-space/vpn')
       .reply(202)
-    return cmd.run({flags: {space: 'my-space', confirm: 'my-space'}})
+    return cmd.run({args: {}, flags: {space: 'my-space', confirm: 'my-space'}})
       .then(() => expect(cli.stderr).to.equal(
-        `Tearing down VPN in space my-space... done\n`))
+        `Tearing down VPN Connection in space my-space... done\n`))
       .then(() => api.done())
   })
 })