+++ packages/spaces/test/commands/vpn/wait.js
@@ -16,9 +16,13 @@ describe('spaces:vpn:wait', function () {
       .get('/spaces/my-space/vpn')
       .reply(200, {
         id: '123456789012',
+        name: 'vpn-connection-name',
         public_ip: '35.161.69.30',
         routable_cidrs: [ '172.16.0.0/16' ],
-        state: 'pending',
+        ike_version: 1,
+        space_cidr_block: '10.0.0.0/16',
+        status: 'pending',
+        status_message: '',
         tunnels: [
           {
             last_status_change: '2016-10-25T22:09:05Z',
@@ -37,19 +41,23 @@ describe('spaces:vpn:wait', function () {
       .get('/spaces/my-space/vpn')
       .reply(200, {
         id: '123456789012',
+        name: 'vpn-connection-name',
         public_ip: '35.161.69.30',
         routable_cidrs: [ '172.16.0.0/16' ],
-        state: 'available',
+        ike_version: 1,
+        space_cidr_block: '10.0.0.0/16',
+        status: 'active',
+        status_message: '',
         tunnels: [
           {
             last_status_change: '2016-10-25T22:09:05Z',
-            outside_ip_address: '52.44.146.197',
+            ip: '52.44.146.197',
             status: 'UP',
             status_message: 'status message'
           },
           {
             last_status_change: '2016-10-25T22:09:05Z',
-            outside_ip_address: '52.44.146.197',
+            ip: '52.44.146.197',
             status: 'UP',
             status_message: 'status message'
           }
@@ -85,9 +93,9 @@ describe('spaces:vpn:wait', function () {
         ike_version: 1
       })
 
-    return cmd.run({flags: {space: 'my-space', interval: 0}})
+    return cmd.run({flags: {space: 'my-space', name: 'vpn-connection-name', interval: 0}})
       .then(() => expect(cli.stderr).to.equal(
-        `Waiting for VPN in space my-space to allocate... done\n\n`))
+        `Waiting for VPN Connection vpn-connection-name to allocate... done\n\n`))
       .then(() => expect(cli.stdout).to.equal(
         `=== my-space VPNs
 VPN Tunnel  Customer Gateway  VPN Gateway    Pre-shared Key  Routable Subnets  IKE Version
@@ -105,6 +113,7 @@ Tunnel 2    52.44.146.196     52.44.146.196  apresharedkey2  10.0.0.0/16       1
       .get('/spaces/my-space/vpn')
       .reply(200, {
         id: '123456789012',
+        name: 'vpn-connection-name',
         public_ip: '35.161.69.30',
         routable_cidrs: [ '172.16.0.0/16' ],
         state: 'pending',
@@ -126,6 +135,7 @@ Tunnel 2    52.44.146.196     52.44.146.196  apresharedkey2  10.0.0.0/16       1
       .get('/spaces/my-space/vpn')
       .reply(200, {
         id: '123456789012',
+        name: 'vpn-connection-name',
         public_ip: '35.161.69.30',
         routable_cidrs: [ '172.16.0.0/16' ],
         state: 'deleted',
@@ -147,7 +157,7 @@ Tunnel 2    52.44.146.196     52.44.146.196  apresharedkey2  10.0.0.0/16       1
         ]
       })
 
-    return cmd.run({flags: {space: 'my-space', interval: 0}})
+    return cmd.run({flags: {space: 'my-space', name: 'vpn-connection-name', interval: 0}})
       .catch(reason => {
         expect(reason.message).to.equal('supplied CIDR block already in use')
       })